server/requirements.txt
alembic==1.13.0
aniso8601==9.0.1
asttokens==2.4.1
backcall==0.2.0
certifi==2023.11.17
click==8.1.7
cloudinary==1.37.0
decorator==5.1.1
executing==2.0.1
Faker==21.0.0
Flask==2.2.5
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
flask-swagger==0.2.14
greenlet==3.0.2
gunicorn==21.2.0
importlib-metadata==7.0.0
importlib-resources==6.1.1
ipdb==0.13.9
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.23
SQLAlchemy-serializer==1.4.1
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==2.2.2
WTForms==3.1.1
zipp==3.17.0
